FPN Grace environment fixes

+++ models/FPN/train/train.py
@@ -297,7 +297,7 @@ def train_one_epoch(sess, ops, train_writer):
     # Shuffle train samples
     train_idxs = np.arange(0, len(TRAIN_DATASET))
     np.random.shuffle(train_idxs)
-    num_batches = len(TRAIN_DATASET) / BATCH_SIZE
+    num_batches = int(len(TRAIN_DATASET) / BATCH_SIZE)
 
     # To collect statistics
     total_correct = 0
@@ -399,7 +399,7 @@ def eval_one_epoch(sess, ops, test_writer):
     log_string(str(datetime.now()))
     log_string("---- EPOCH %03d EVALUATION ----" % (EPOCH_CNT))
     test_idxs = np.arange(0, len(TEST_DATASET))
-    num_batches = len(TEST_DATASET) / BATCH_SIZE
+    num_batches = int(len(TEST_DATASET) / BATCH_SIZE)
 
     # To collect statistics
     total_correct = 0
